Split files into logical folders
