+++ src/models/instruction.js
@@ -1,0 +1,8 @@
+export function Instruction(instruction) {
+    const trimmed = instruction.trim();
+
+    if (trimmed === '') {
+        throw new Error(`Invalid instruction ${instruction}`);
+    }
+    this.instruction = trimmed;
+}
