+++ src/parsers/stitch-parser.js
