+++ src/utils/round-resolvers.js
@@ -1,4 +1,4 @@
-import { FlatStitchArrayGenerator } from './stitch.js';
+import { FlatStitchArrayGenerator } from "./flat-stitch-array-generator";
 import { StitchCounter } from './stitch-counter.js';
 
 export function RoundCountResolver(flattener = FlatStitchArrayGenerator.fromArray, counter = StitchCounter.getCount) {
@@ -18,19 +18,3 @@ export function RoundStitchIndexResolver(flattener = FlatStitchArrayGenerator.fr
     }
 }
 
-export function Round(stitches, instruction) {
-    if (!stitches || stitches.length === 0) {
-        throw new Error('Empty round');
-    }
-
-    this._stitches = stitches;
-    this._instruction = instruction;
-}
-Object.assign(Round.prototype, {
-    getStitchCount: function(resolver = RoundCountResolver()) {
-        return resolver(this._stitches);
-    },
-    getStitchAtIndex: function(index, resolver = RoundStitchIndexResolver()) {
-        return resolver(index, this._stitches);
-    },
-})
