+++ src/utils/stitch-counter.js
